+++ src/Page/Admin/Modal/UpdateModalChildren.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useRecoilState, useSetRecoilState } from "recoil";
 import styled from "styled-components";
@@ -7,18 +7,25 @@ import {
   productListState,
   selectOptionState,
   selectProductListState,
+  Option,
 } from "../../../state/productItemState";
-import { SelectOption } from "../../../Page/Admin/AdminManageProductItemList";
+
 import UpdateModalForm from "../UpdateModalForm";
 
 const Wrapper = styled.div`
   display: grid;
-  grid-template-rows: 1fr 27rem 0.5fr;
+  overflow-y: hidden;
+  grid-template-rows: 1fr 5fr 0.5fr;
 `;
 
-const StateInfoContainer = styled.div``;
+const StateInfoContainer = styled.div`
+  margin-bottom: 1rem;
+`;
 
 const ButtonContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
   button {
     cursor: pointer;
     font-size: 2rem;
@@ -26,16 +33,12 @@ const ButtonContainer = styled.div`
     padding: 0.8rem 1.3rem;
     border-radius: 0.5rem;
     color: ${(props) => props.theme.color.fontColorWhite};
-    &:nth-child(1) {
+    &.cancel-button {
       background-color: ${(props) => props.theme.color.gray200};
+      margin-right: 0.8rem;
     }
-
-    &:nth-child(2) {
-      background-color: ${(props) => props.theme.color.error500};
-    }
-
-    &:not(:first-child) {
-      margin-left: 0.8rem;
+    &.confirm-button {
+      background-color: ${(props) => props.theme.color.secondary500};
     }
   }
 `;
@@ -45,26 +48,18 @@ const FormContainer = styled.div`
   overflow-y: auto;
   form {
     box-sizing: border-box;
-    .product-form-container {
-      box-sizing: border-box;
-      padding: 1rem 1rem;
-      border: 1px solid ${(props) => props.theme.color.gray200};
-      display: flex;
-      flex-direction: column;
-      margin-bottom: 1rem;
-    }
   }
 `;
 
-interface ISelectModalChildrenProps<T> {
+interface ISelectModalChildrenProps {
   setModal: React.Dispatch<React.SetStateAction<boolean>>;
-  items?: Array<T>;
+  items?: ProductListValues[];
 }
 
 const UpdateModalChildren = ({
   setModal,
   items,
-}: ISelectModalChildrenProps<ProductListValues>) => {
+}: ISelectModalChildrenProps) => {
   const setSelectOption = useSetRecoilState(selectOptionState);
   const [productList, setProductList] = useRecoilState(productListState);
   const [selectList, setSelectList] = useRecoilState<ProductListValues[]>(
@@ -78,11 +73,12 @@ const UpdateModalChildren = ({
       ...preValue.map((value) => ({ ...value, select: false })),
     ]);
     setModal(false);
-    setSelectOption({ option: "none" });
+    setSelectOption({ options: Option.NONE });
   };
 
   const selectUpdateItemsSubmitHandler = handleSubmit((data: object) => {
     const dataArray = Object.entries(data);
+
     let selectValue: {
       name: string;
       desc: string;
@@ -98,28 +94,29 @@ const UpdateModalChildren = ({
         return Number(key) === product.id;
       });
 
-      if (selectProduct) {
-        if (selectValue.option !== "") {
-          selectValue.option = selectValue.option.split(",");
-        } else {
-          selectValue.option = [];
-        }
-
-        if (selectValue.thumbnail.length === 0) {
-          selectValue.thumbnail = null;
-        } else {
-          selectValue.thumbnail = selectValue.thumbnail[0].name;
-        }
-
-        return { ...product, ...selectValue, select: false };
+      if (!selectProduct) {
+        return product;
       }
-      return product;
+
+      if (selectValue.option !== "") {
+        selectValue.option = selectValue.option.split(",");
+      } else {
+        selectValue.option = [];
+      }
+
+      if (selectValue.thumbnail.length === 0) {
+        selectValue.thumbnail = null;
+      } else {
+        selectValue.thumbnail = selectValue.thumbnail[0].name;
+      }
+
+      return { ...product, ...selectValue, select: false };
     });
 
     setProductList(newProductList);
     setModal(false);
     setSelectList([]);
-    setSelectOption({ option: "none" });
+    setSelectOption({ options: Option.NONE });
   });
 
   useEffect(() => {
@@ -132,46 +129,19 @@ const UpdateModalChildren = ({
     }
   }, []);
 
-  /* 
-  form 필드 선택시 유저가 입력하려는 필드가 선택됐다는 것을 표시하기 위한 코드지만 
-  useRef를 이용해서 조금 더 손쉽게 수정이 가능한 것으로 보이기 때문에 
-  먼저 수정 데이터를 처리하는 로직이 끝난 다음에 바로 수정할 해야한다!!!!!
-  
+  /*
+  TODO: form 필드 선택시 백그라운드를 변경하는 로직 필요
   */
-  // const [currentSelectFormContainer, setCurrentSelectFormContainer] =
-  //   useState("");
-
-  // const selectFormContainer = (e: React.MouseEvent<HTMLFieldSetElement>) => {
-  //   const id = e.currentTarget.dataset.id;
-  //   if (id) {
-  //     setCurrentSelectFormContainer(id);
-  //   }
-  // };
-
-  // useEffect(() => {
-  //   if (currentSelectFormContainer) {
-  //     const containerList = document.querySelectorAll<HTMLDivElement>(
-  //       ".product-form-container",
-  //     );
-  //     if (containerList) {
-  //       const [selected] = Array.from(containerList)
-  //         .map((item) => {
-  //           item.style.backgroundColor = "white";
-  //           return item;
-  //         })
-  //         .filter((value) => value.dataset.id === currentSelectFormContainer);
-  //       selected.style.backgroundColor = "#E0FFDF";
-  //     }
-  //   }
-  // }, [currentSelectFormContainer]);
 
   return (
     <>
       <Wrapper>
         <StateInfoContainer>
           <h1>수정하기</h1>
-          <h2>수정할 내용을 입력하고 버튼을 누르세요.</h2>
-          <p>수정 할 내용을 입력하지 않으면 이전 내용으로 저장됩니다.</p>
+          <p>
+            수정할 내용을 입력하고 버튼을 누르세요. 수정 할 내용을 입력하지
+            않으면 이전 내용으로 저장됩니다.
+          </p>
         </StateInfoContainer>
         <FormContainer>
           <form>
@@ -179,7 +149,7 @@ const UpdateModalChildren = ({
               const fieldName = item.id;
               return (
                 <UpdateModalForm
-                  fieldName={fieldName}
+                  fieldName={fieldName.toString()}
                   register={register}
                   item={item}
                 />
@@ -188,8 +158,15 @@ const UpdateModalChildren = ({
           </form>
         </FormContainer>
         <ButtonContainer>
-          <button onClick={handleModal}>돌아가기</button>
-          <button onClick={selectUpdateItemsSubmitHandler}>수정하기</button>
+          <button className="cancel-button" onClick={handleModal}>
+            돌아가기
+          </button>
+          <button
+            className="confirm-button"
+            onClick={selectUpdateItemsSubmitHandler}
+          >
+            수정하기
+          </button>
         </ButtonContainer>
       </Wrapper>
     </>
